backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
passlib[argon2]==1.7.4
requests==2.31.0
dnspython==2.4.2

